Add --noauto (disabled at the moment) for automatic loading of base packages
Normally when the hl building changes are in we want to load
base and haskell98 hl files unless the user says --noauto.

This patch implements --noauto with the exception that
it is the default. The default should be flipped after
the library changes are committed.

darcs-hash:20050923110453-e6152-1136862861c43e1364698317335782d3299a1adc.gz